calculate_statistics/requirements.txt
performanceplatform-client>=0.6.0